+++ DirectX 11 Playground/Src/Graphics/Graphics.cpp
@@ -428,8 +428,8 @@ bool Graphics::initDirectX(HWND hwnd, int width, int height)
 
 		viewport.TopLeftX = 0;
 		viewport.TopLeftY = 0;
-		viewport.Width = width;
-		viewport.Height = height;
+		viewport.Width = static_cast<float>(width);
+		viewport.Height = static_cast<float>(height);
 		viewport.MinDepth = 0.0f;
 		viewport.MaxDepth = 1.0f;
 		//Set the vieport
